refactor(imports): centralize the import sort key and same-group predicate
- Relocate `import_sort_key` and `least_alias` with their tests from `alphabetize` into `primitives::imports`.
- Extract `same_import_group`, routing `banded_gap` and the import-run gap filter through one predicate.
