+++ src/rules/alphabetize/mod.rs
@@ -21,8 +21,7 @@ use std::{borrow::Cow, collections::HashMap, ops::Range};
 
 use ruff_diagnostics::Edit;
 use ruff_python_ast::{
-    Alias, Decorator, ExceptHandler, Expr, ExprDict, PythonVersion, Stmt, StmtAnnAssign,
-    StmtFunctionDef,
+    Decorator, ExceptHandler, Expr, ExprDict, PythonVersion, Stmt, StmtAnnAssign, StmtFunctionDef,
     helpers::{any_over_expr, is_compound_statement, is_dunder, map_callable},
 };
 use ruff_source_file::LineRanges;
@@ -32,7 +31,7 @@ use crate::{
     config::Config,
     primitives::{
         edit::{apply_inline_edits, narrowed_replacement, singleton_groups},
-        imports::{ImportGroup, future_annotations_alias, import_group},
+        imports::{future_annotations_alias, import_sort_key, same_import_group},
         orderer::{assemble_blocks, block_range, blocks_span, permute_full, permute_in_place},
         params::pins_positional_params,
         scope::{BodyScope, scoped_body},
@@ -254,33 +253,6 @@ fn has_keep_marker(source: &Source, dict: &ExprDict) -> bool {
         .any(|c| source.slice(c).trim_start_matches('#').trim() == "prose: keep")
 }
 
-/// Composite import sort key landing the canonical group order
-/// (bare → external `from` → local-package) ahead of a per-kind
-/// inner sort. Within a group, bare imports sort before `from`
-/// imports, bare by least alias name and `from` by `(level, module)`.
-/// `None` pins any non-import statement in place.
-fn import_sort_key<'a>(
-    stmt: &'a Stmt,
-    first_party: &[String],
-) -> Option<(ImportGroup, u8, u32, &'a str)> {
-    let group = import_group(stmt, first_party)?;
-    Some(match stmt {
-        Stmt::Import(i) => (group, 0, 0, least_alias(&i.names)),
-        Stmt::ImportFrom(i) => (group, 1, i.level, i.module.as_deref().unwrap_or_default()),
-        _ => unreachable!("import_group returns Some only for import statements"),
-    })
-}
-
-/// Returns the alphabetically least alias name in a bare import's
-/// name list. An `import` statement always binds at least one name.
-fn least_alias(names: &[Alias]) -> &str {
-    names
-        .iter()
-        .map(|a| a.name.as_str())
-        .min()
-        .expect("import binds at least one name")
-}
-
 /// Returns the method-group index. `0` for dunders, `1` for
 /// `@property` / `@cached_property` (decided by the first decorator),
 /// `2` for single-leading-underscore privates, `3` for public.
@@ -380,11 +352,9 @@ fn rewrite_body<'a>(
         // same-group import neighbors collapse to one line, derived from the
         // assembled order the family sorts left.
         if band_ranks.is_none() {
-            let group = |slot: usize| import_group(&body[order[slot]], first_party);
-            import_run_slots.extend(
-                (0..n.saturating_sub(1))
-                    .filter(|&slot| group(slot).is_some() && group(slot) == group(slot + 1)),
-            );
+            import_run_slots.extend((0..n.saturating_sub(1)).filter(|&slot| {
+                same_import_group(&body[order[slot]], &body[order[slot + 1]], first_party)
+            }));
         }
     }
     let any_owned = rendered.iter().any(|c| matches!(c, Cow::Owned(_)));
@@ -584,35 +554,6 @@ mod tests {
             has_inline_statement_join(&source, &source.ast().body),
             expected
         );
-    }
-
-    #[test]
-    fn import_sort_key_ranks_groups_then_bare_before_from_within_local() {
-        let first_party = vec!["myapp".to_owned()];
-        let s = parse("import os\nfrom os import path\nimport myapp.core\nfrom myapp import app\n");
-        let keys: Vec<_> = s
-            .ast()
-            .body
-            .iter()
-            .map(|stmt| import_sort_key(stmt, &first_party).expect("import statement"))
-            .collect();
-        assert!(
-            keys[0] < keys[1] && keys[1] < keys[2] && keys[2] < keys[3],
-            "expected bare-external < external-from < local-bare < local-from",
-        );
-    }
-
-    #[test]
-    fn import_sort_key_returns_none_for_non_import() {
-        let s = parse("x = 1\n");
-        assert!(import_sort_key(&s.ast().body[0], &[]).is_none());
-    }
-
-    #[test]
-    fn least_alias_returns_alphabetically_min_name() {
-        let s = parse("import sys, os, abc\n");
-        let import = s.ast().body[0].as_import_stmt().expect("import");
-        assert_eq!(least_alias(&import.names), "abc");
     }
 
     #[test]
